PM-3323: Split out test composition.
