+++ metronome/examples/specs/src/io/iohk/metronome/examples/robot/app/RobotCompositionSpec.scala
@@ -1,32 +1,18 @@
 package io.iohk.metronome.examples.robot.app
 
 import cats.implicits._
-import cats.effect.{Blocker, Resource}
-import cats.effect.concurrent.Ref
-import io.iohk.metronome.crypto.{ECKeyPair, ECPublicKey}
-import io.iohk.metronome.networking.{
-  RemoteConnectionManager,
-  ConnectionHandler,
-  NetworkTracers
-}
-import io.iohk.metronome.hotstuff.service.tracing.{
-  ConsensusEvent,
-  ConsensusTracers,
-  SyncTracers
-}
-import io.iohk.metronome.examples.robot.app.tracing._
+import cats.effect.Resource
+import io.iohk.metronome.crypto.ECKeyPair
+import io.iohk.metronome.hotstuff.service.tracing.ConsensusEvent
 import io.iohk.metronome.examples.robot.app.config.{
   RobotConfig,
   RobotConfigParser,
   RobotOptions
 }
-import io.iohk.metronome.logging.{InMemoryLogTracer, HybridLog, HybridLogObject}
-import io.iohk.metronome.tracer.Tracer
+import io.iohk.metronome.logging.HybridLogObject
 import java.nio.file.Files
-import java.net.InetSocketAddress
 import monix.eval.Task
 import monix.execution.schedulers.TestScheduler
-import monix.tail.Iterant
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.compatible.Assertion
 import scala.concurrent.duration._
@@ -71,7 +57,7 @@ class RobotCompositionSpec extends AnyFlatSpec with Matchers {
   it should "compose components that can run and stay in sync" in test {
     new Fixture(10.minutes) {
       // Wait with the test result to keep the resources working.
-      override def test(envs: List[Env]) =
+      override def test(envs: List[RobotTestComposition.Env]) =
         for {
           _               <- Task.sleep(duration - 1.minute)
           logs            <- envs.traverse(_.logTracer.getLogs)
@@ -90,39 +76,12 @@ class RobotCompositionSpec extends AnyFlatSpec with Matchers {
 }
 
 object RobotCompositionSpec {
-  import RobotConsensusTracers.RobotConsensusEvent
-  import RobotSyncTracers.RobotSyncEvent
-
-  class EventTracer[A](eventLogRef: Ref[Task, Vector[A]])
-      extends Tracer[Task, A] {
-
-    override def apply(a: => A): Task[Unit] =
-      eventLogRef.update(_ :+ a)
-
-    val clear     = eventLogRef.set(Vector.empty)
-    val getEvents = eventLogRef.get
-  }
-  object EventTracer {
-    def apply[A] =
-      new EventTracer[A](Ref.unsafe[Task, Vector[A]](Vector.empty))
-  }
-
-  /** Things we may want to access in tests. */
-  case class Env(
-      storages: TestComposition#Storages,
-      logTracer: InMemoryLogTracer.HybridLogTracer[Task],
-      consensusEventTracer: EventTracer[RobotConsensusEvent],
-      syncEventTracer: EventTracer[RobotSyncEvent]
-  ) {
-    val clear =
-      logTracer.clear >> consensusEventTracer.clear >> syncEventTracer.clear
-  }
 
   abstract class Fixture(val duration: FiniteDuration)
       extends RobotComposition {
 
     /** Override to implement the test. */
-    def test(envs: List[Env]): Task[Assertion]
+    def test(envs: List[RobotTestComposition.Env]): Task[Assertion]
 
     val scheduler = TestScheduler()
 
@@ -169,95 +128,11 @@ object RobotCompositionSpec {
         nodeEnvs <- (0 until config.network.nodes.size).toList.map { i =>
           val opts = RobotOptions(nodeIndex = i)
           val comp = makeComposition(scheduler)
-          comp.compose(opts, config).map { storages =>
-            Env(
-              storages,
-              comp.logTracer,
-              comp.consensusEventTracer,
-              comp.syncEventTracer
-            )
-          }
+          comp.composeEnv(opts, config)
         }.sequence
-
-        _ <- Resource.pure[Task, Unit](()).onFinalize {
-          nodeEnvs.traverse(_.clear).void
-        }
       } yield nodeEnvs
 
     def makeComposition(scheduler: TestScheduler) =
-      new TestComposition(scheduler)
+      new RobotTestComposition(scheduler)
   }
-
-  // Every node has its own composer, so we can track logs separately.
-  class TestComposition(scheduler: TestScheduler) extends RobotComposition {
-
-    /** Class level log collector.
-      *
-      * If the composer is reused, this should be cleared between tests.
-      */
-    val logTracer            = InMemoryLogTracer.hybrid[Task]
-    val consensusEventTracer = EventTracer[RobotConsensusEvent]
-    val syncEventTracer      = EventTracer[RobotSyncEvent]
-
-    private def makeLogTracer[T: HybridLog] =
-      InMemoryLogTracer.hybrid[Task, T](logTracer)
-
-    override protected def makeNetworkTracers = {
-      import RobotNetworkTracers._
-      NetworkTracers(makeLogTracer[RobotNetworkEvent])
-    }
-
-    override protected def makeConsensusTracers = {
-      import RobotConsensusTracers._
-      ConsensusTracers(
-        makeLogTracer[RobotConsensusEvent] |+| consensusEventTracer
-      )
-    }
-
-    override protected def makeSyncTracers = {
-      import RobotSyncTracers._
-      SyncTracers(makeLogTracer[RobotSyncEvent] |+| syncEventTracer)
-    }
-
-    // Use the `TestScheduler` to block on queries, otherwise the test hangs.
-    override protected def makeDBBlocker =
-      Resource.pure[Task, Blocker](Blocker.liftExecutionContext(scheduler))
-
-    // TODO: Simulate a network.
-    // NOTE: We cannot use a real network with the `TestScheduler`.
-    override protected def makeConnectionManager(
-        config: RobotConfig,
-        opts: RobotOptions
-    )(implicit
-        networkTracers: NTS
-    ) = Resource.pure[Task, ConnectionManager] {
-      val localNode = config.network.nodes(opts.nodeIndex)
-
-      new RemoteConnectionManager[Task, ECPublicKey, NetworkMessage] {
-        override val getLocalPeerInfo: (ECPublicKey, InetSocketAddress) =
-          (
-            localNode.publicKey,
-            new InetSocketAddress(localNode.host, localNode.port)
-          )
-
-        override def getAcquiredConnections: Task[Set[ECPublicKey]] = Task {
-          config.network.nodes.map(_.publicKey).toSet - localNode.publicKey
-        }
-
-        override def incomingMessages: Iterant[
-          Task,
-          ConnectionHandler.MessageReceived[ECPublicKey, NetworkMessage]
-        ] = Iterant.never
-
-        override def sendMessage(
-            recipient: ECPublicKey,
-            message: NetworkMessage
-        ): Task[Either[ConnectionHandler.ConnectionAlreadyClosedException[
-          ECPublicKey
-        ], Unit]] = Task.now(Right(()))
-
-      }
-    }
-  }
-
 }
